Consolidate parameter file write to P_tmpdir
* Move functionality into KIM_SharedLibrary as convenience routines
* Update KIM_SimulatorModelImplementation to use new routines
* Update KIM_ModelImplementation to use new routines

Changes are backward compatible
